// src/utils/interface.ts
import { TextStyle, ViewStyle } from "react-native";

export interface Styles {
    container: ViewStyle;
    header: ViewStyle;
    title: TextStyle;
    filterButton: ViewStyle;
    listContent: ViewStyle;
    notificationItem: ViewStyle;
    unreadNotificationItem: ViewStyle;
    iconContainer: ViewStyle;
    notificationContent: ViewStyle;
    notificationHeader: ViewStyle;
    notificationTitle: TextStyle;
    unreadNotificationTitle: TextStyle;
    notificationMessage: TextStyle;
    notificationDate: TextStyle;
    separator: ViewStyle;
    matchIcon: ViewStyle;
    reminderIcon: ViewStyle;
    paymentIcon: ViewStyle;
    validationIcon: ViewStyle;
    cancelIcon: ViewStyle;
    creditIcon: ViewStyle;
    infoIcon: ViewStyle;
    emailIcon: ViewStyle;
    smsIcon: ViewStyle;
    pushIcon: ViewStyle;
    unreadIndicator: ViewStyle;
}